Moved atmosphere.py in fastoad.base
