+++ src/fastoad/base/atmosphere.py
