+++ src/fastoad/model_base/propulsion.py
@@ -22,7 +22,7 @@ import pandas as pd
 from openmdao import api as om
 from openmdao.core.component import Component
 
-from fastoad.base.flight_point import FlightPoint
+from fastoad.base import FlightPoint
 
 
 class IPropulsion(ABC):
