+++ src/fastoad/models/handling_qualities/tail_sizing/compute_ht_area.py
@@ -2,7 +2,7 @@
 Estimation of horizontal tail area
 """
 #  This file is part of FAST-OAD : A framework for rapid Overall Aircraft Design
-#  Copyright (C) 2020  ONERA & ISAE-SUPAERO
+#  Copyright (C) 2021 ONERA & ISAE-SUPAERO
 #  FAST is free software: you can redistribute it and/or modify
 #  it under the terms of the GNU General Public License as published by
 #  the Free Software Foundation, either version 3 of the License, or
@@ -18,7 +18,7 @@ import numpy as np
 import openmdao.api as om
 from scipy.constants import g
 
-from fastoad.utils.physics import Atmosphere
+from fastoad.base import Atmosphere
 
 
 class ComputeHTArea(om.ExplicitComponent):
