+++ src/fastoad/models/performances/mission/base.py
@@ -1,6 +1,6 @@
 """Base classes for mission computation."""
 #  This file is part of FAST-OAD : A framework for rapid Overall Aircraft Design
-#  Copyright (C) 2020  ONERA & ISAE-SUPAERO
+#  Copyright (C) 2021 ONERA & ISAE-SUPAERO
 #  FAST is free software: you can redistribute it and/or modify
 #  it under the terms of the GNU General Public License as published by
 #  the Free Software Foundation, either version 3 of the License, or
@@ -18,7 +18,7 @@ from typing import List
 
 import pandas as pd
 
-from fastoad.base.flight_point import FlightPoint
+from fastoad.base import FlightPoint
 
 
 class IFlightPart(ABC):
