+++ src/fastoad/models/performances/mission/segments/base.py
@@ -22,10 +22,9 @@ import pandas as pd
 from scipy.constants import g
 from scipy.optimize import root_scalar
 
-from fastoad.base.flight_point import FlightPoint
+from fastoad.base import AtmosphereSI, FlightPoint
 from fastoad.constants import EngineSetting
 from fastoad.model_base.propulsion import IPropulsion
-from fastoad.utils.physics import AtmosphereSI
 from ..base import IFlightPart
 from ..exceptions import FastFlightSegmentIncompleteFlightPoint
 from ..polar import Polar
