+++ src/fastoad/models/performances/mission/segments/tests/test_segments.py
@@ -17,14 +17,14 @@ import pytest
 from numpy.testing import assert_allclose
 from scipy.constants import foot
 
-from fastoad.base.flight_point import FlightPoint
+from fastoad.base import FlightPoint
 from fastoad.model_base.propulsion import AbstractFuelPropulsion, FuelEngineSet
 from ..altitude_change import AltitudeChangeSegment
 from ..cruise import (
-    OptimalCruiseSegment,
-    CruiseSegment,
     BreguetCruiseSegment,
     ClimbAndCruiseSegment,
+    CruiseSegment,
+    OptimalCruiseSegment,
 )
 from ..hold import HoldSegment
 from ..speed_change import SpeedChangeSegment
