+++ src/fastoad/models/performances/mission/segments/transition.py
@@ -14,11 +14,11 @@
 
 from copy import deepcopy
 from dataclasses import dataclass
-from typing import Tuple, List
+from typing import List, Tuple
 
 import pandas as pd
 
-from fastoad.base.flight_point import FlightPoint
+from fastoad.base import FlightPoint
 from fastoad.model_base.propulsion import IPropulsion
 from fastoad.models.performances.mission.polar import Polar
 from fastoad.models.performances.mission.segments.base import FlightSegment
