+++ src/fastoad/models/performances/mission/tests/test_routes.py
@@ -32,11 +32,11 @@ import pandas as pd
 import pytest
 from matplotlib.ticker import MultipleLocator
 from numpy.testing import assert_allclose
-from scipy.constants import knot, foot
+from scipy.constants import foot, knot
 
-from fastoad.base.flight_point import FlightPoint
-from fastoad.constants import FlightPhase, EngineSetting
-from fastoad.model_base.propulsion import IPropulsion, FuelEngineSet
+from fastoad.base import FlightPoint
+from fastoad.constants import EngineSetting, FlightPhase
+from fastoad.model_base.propulsion import FuelEngineSet, IPropulsion
 from fastoad.models.performances.mission.base import IFlightPart
 from fastoad.models.performances.mission.polar import Polar
 from fastoad.models.performances.mission.routes import RangedRoute
